+++ src/sim/meca1d.h
@@ -140,7 +140,7 @@ public:
     real setRightHandSide(real kT)
     {
         real res = INFINITY;
-        for ( unsigned ii = 0; ii < mecables.size(); ++ii )
+        for ( size_t ii = 0; ii < mecables.size(); ++ii )
         {
             real b = sqrt( 2 * kT * vMOB[ii] );
             vRHS[ii] = vMOB[ii] * vBAS[ii] + b * RNG.gauss();
